\name{apt-package}
\alias{apt}
\alias{apt-package}
\docType{package}
\title{Asymmetric Price Transmission}

\description{
Asymmetric price transmission between two time series is assessed. The names of functions and datasets reveal the categories they belong to. A prefix of \code{da} is for datasets, \code{ci} for cointegration, and \code{ecm} for error correction model.

The focus is on the price transmission between \emph{two} price variables. Therefore, objectives like fitting an error correction model for more than two variables are beyond the scope of this package.
}

\details{
\tabular{ll}{
Package: \tab apt\cr
Type: \tab Package\cr
Version: \tab 3.0\cr
Date: \tab 2020-04-30\cr
Depends: \tab R (>= 3.0.0), erer \cr
Imports: \tab car, urca \cr
License: \tab GPL \cr
LazyLoad: \tab yes\cr
}
}

\author{Changyou Sun (\email{cs258@msstate.edu})}
\keyword{package} 