\name{apt-package}
\alias{apt}
\alias{apt-package}
\docType{package}
\title{Asymmetric Price Transmission}
\description{
This package focuses on asymmetric price transmission between two time series. The name of functions and datasets reveals the categories they belong to. A prefix of \code{da} is for datasets, \code{ci} for cointegration, and \code{ecm} for error correction model.

This package focuses on the price transmission between \emph{two} price variables. Therefore, objectives like fitting an error correction model for more than two variables are beyond the scope of this package.
}
\details{
\tabular{ll}{
Package: \tab apt\cr
Type: \tab Package\cr
Version: \tab 2.1\cr
Date: \tab 2010-11-03 (first built); 2014-6-3 (last) \cr
Depends: \tab R (>= 3.0.0), erer, urca \cr
Imports: car \cr
License: \tab GPL \cr
LazyLoad: \tab yes\cr
}
}
\author{
Changyou Sun <csun@cfr.msstate.edu>
}
\keyword{package} 